+++ app/deeplio/Deeplio.php
@@ -77,6 +77,9 @@ namespace noelbosscom;
 			if(file_exists($path.'.config.json')){
 
 				$conf = json_decode( file_get_contents( $path.'.config.json' ) );
+				if(!$conf->enabled){
+					$this->log('[NOTE] Repository disabled by config', true);
+				}
 				$this->projectconf = $conf;
 
 				if($conf === null || !is_object($conf)){