+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added enabled flag to config
 * Checks if mails are configured
 * Mails are now arrays with Names and E-Mails
+* Updated Log Information
 
 ## v0.3.3 - 2015-06-23